Fix burrower lifeboat tunneling (#5778)
# About the pull request

This PR effectively revives the fix from #5603 that was allowing
burrowing in lifeboats. The other grammar changes it was doing were for
marines, so it would not be correct to we/our them.

# Explain why it's good for the game

Fixes burrowing into lifeboats.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl: AtticusRezzer
fix: Fixed the incorrect area flag in lifeboats allowing tunneling
/:cl: