MOPP Desc tism (#5782)
# About the pull request
Shortens the active description for the MOPP suit so it isn't a bible's
worth of text explaining how a MOPP suit works that extends further then
the lore desc. This should be a very simple merge.


# Explain why it's good for the game

A general item desc shouldn't be excessively long.



# Changelog
:cl:
del: Removes part of MOPP suit desc.
/:cl: