Refactor marine_armor.dm and split it into multiples files. (#5740)
# About the pull request
1-Refactor the marine_armor.dm file by splitting them into multiples
files.
2-in marine_armor.dm i also remove a serie of duplicate declaration of
the same items.
<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
refactor: refactor marine_armor.dm and split it into multiples files.
/:cl:

---------

Co-authored-by: Julien <jverger.ingx@gmail.com>
Co-authored-by: SabreML <57483089+SabreML@users.noreply.github.com>