Barricades Multi-Weld Revert & Refactoring (#5795)
# About the pull request

This PR is a follow up to #5609 which removed a busy check from some
barricades allowing multi-welding. This PR restores that setting (which
is a return to normal for all but folding barricades which incorrectly
allowed multi-welding before).

I have also added automatic looping to barricade welding if there is
still work to be done, and it can be performed.

# Explain why it's good for the game

The PR that introduced this change was not intended to be a balance
change. Additionally, it's far more difficult to tune welding speeds if
the quantity of barricades being welded at the same time is variable.
Technically this is a nerf though because previously multi-welding
folding barricades was possible.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>


https://github.com/cmss13-devs/cmss13/assets/76988376/32deb97f-31b0-4fd0-8baf-54637cfbf81d

After looping welding addition:


https://github.com/cmss13-devs/cmss13/assets/76988376/e67efe06-711e-4847-aa36-345c1e85222d

</details>


# Changelog
:cl: Drathek
add: Welding a barricade will now repeat automatically if it can still
be repaired.
balance: Reverted multi-welding change to barricades: All barricades
must be welded one at a time.
refactor: Refactored more barricade code.
/:cl: