Tweaks oppressor (#5759)
# About the pull request

Raises the tailhook from 7 deciseconds to 8 deciseconds.

# Explain why it's good for the game

On my last PR i tweaked this from 8 to 7 thinking it wouldnt be too
impactful but people giving me ""FEEDBACK"" have made me realise this is
a bit too fast with players with 200+ ping above, originally i wasnt
going to do this because i was planning on a full rework but i dont want
to leave it ''broken'' for the time being.


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
balance: Oppressor tailhook is now 8 deciseconds instead of 7.
/:cl:

Co-authored-by: InsaneRed <prodexter31@outlook.comm>