Shiva's Panic Room Nightmare Insert (#5800)
# About the pull request

On the day of the xenomorph incursion on the colony, the alarms started
blaring calling all colonists to take shelter and hide, top
Administrators made a break for the Spaceport shuttles. All but one, the
Assistant Operations Manager knew the shuttles would be gone by now, so
they opened the panic room for all civilians who wanted to be safe from
the aliens.

Now we are in the present day, the supplies are running low, the weapons
are lackluster to deal with the xenos and outside contact has been cut
off, same as the power in some areas. A group of doctors, scientists and
normal civilians lead by the Assistant Manager must find a way to
survive before it is too late for their rescues arrival.

# Explain why it's good for the game

More nightmare inserts are great. I like the story they provide and this
time I did not majorly buff them, hopefully this can be merged same as
the LV one.


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>


![image](https://github.com/cmss13-devs/cmss13/assets/79063506/a7297c7b-4e60-41c6-a336-85dc2ea17d9c)

</details>


# Changelog
:cl:
mapadd: added a new nightmare insert in the Operations Panic Room on
Shivas

---------

Co-authored-by: Jeff Watchson <johncenatoutourourou@gmail.com>