Balance : set base price for B12 to 30 and for the M4 to 20. (#5698)
# About the pull request
1-set base price for B12 to 30 and for the M4 to 20.
2-Medic comtech and SL have a discount of 20% because they have extra
supplies to by that are dependent on their role.
3-I also remove the free M4 for the FTL.
4-change price for drop pouch from 15 to 10

If you think M4 base price is to low i can increase it to 25.

if you think discount is to high or low i can change it...

if you think removing M4 to FTL is to much i can give them a discount
for armors or just for the M4 lowering it to 16 instead of 20

<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
1-base price for B12 was already 30 but i am lowering M4 cost because it
only give limited advantage
2-Medic comtech and SL had all some sort of discount for a bunch of
items so here i am just picking a discount that seem balance.
3-This is to avoid people choosing FTL for the extra free armor instead
of picking Rifleman.
4-I think it's to expensive for what it is and it's very niche use. and
because Spartan asked me.

![image](https://github.com/cmss13-devs/cmss13/assets/117036822/a951f5c7-2593-4e18-bf93-16901dc282b2)

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
balance: set base price for B12 to 30 and for the M4 to 20.
balance: Medic comtech and SL have a discount of 20% for the B12 and M4.
(B12=24 & M4=16=)
balance: removed the free M4 for FTL but added FTL the option to buy it
for 20.
balance: changed the price of Drop Pouch from 15 to 10.
/:cl:

---------

Co-authored-by: Julien <jverger.ingx@gmail.com>