adding vendors for maintenance technician (#5749)
# About the pull request

<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
add: adding vendor for maintenance technician.
code: remove some item that maintenance technician spawned with and put
them in is vendor instead.
maptweak: replaced secured closet by vendor for maintenance technician.
/:cl:

---------

Co-authored-by: Julien <jverger.ingx@gmail.com>