Names for nodes and traps. (#5831)
# About the pull request

Selfevident.

# Explain why it's good for the game

In all honesty, this is a personal gripe. I have poor eyesight and a
laptop with small screen, so this annoys me more than it should.

Xeno traps, by design, look almost like regular weeded tiles, save for
small part of the actual trap in the middle of the weed square. If
properly hidden underneath big enough pile of items, the only way to
tell without rightclicking is to mouseover it and read the name in the
bottom left corner (which is not all that large). With the current
sprite for the weed node being a square of weeds with a small node in
the middle of it, nodes behave exactly the same. That is not a problem
per se: the challenge of telling a trap from a node is no different from
the challenge of telling a trap from regular weeds. The problem lies in
the fact that their supposedly telltale _names_ are the same save for
two letters, and h/n and l/d do not even look all that much different at
a glance.

To make things worse, these two things require exactly opposite
approach. Nodes are completely immune to bullets, and the correct course
of action is to come close and take them out with a bayonet. Traps will
stun you with acid if you come close with a bayonet, and the correct
course of action is to take them out from a distance with a single
bullet. Confusing one for another can be irritating at best and fatal
for your character at worst.

May this little change make life less painful for all the scout players
that ever tried to stab an acid trap.

# Changelog

:cl:
spellcheck: Changed names of "resin node" and "resin hole" to "weed
node" and "resin trap" respectfully.
/:cl: