Faction frequences (#5833)
# About the pull request

I pondered for a bit whether FORECON fixing radiotowers should allow
them to simply use Almayer's frequences (which might require giving them
appropriate headsets or not), but decided to keep it out of scope for
now. Tell me if we actually want it to be a thing.

Also, we apparently have `FACTION_USCM` and `FACTION_MARINE` both
\#defined as same thing. Before it would get out of hand, I am axing it
in favour of more commonly used one of the two.

P.S. I am not sure what is the matter with wiping frequences. Even after
fixing the spelling in the `switch`, it does now remove them from the
list, but that does not seem to prevent them from working. I am calling
that out of scope for now too and leave investigating it for another
day.

# Explain why it's good for the game

Fixes #4766, fixes #5797.

# Changelog

:cl:
fix: FORECON and RMC now can add their frequences to groundside
radiotowers.
/:cl: