Moves current UPP/CLF combat synths into a support role  (#5830)
# About the pull request

We discussed combat synths on the maintainer meeting yesterday,
specifically the UPP/CLF ert's and we've come to the conclusion and
decided we want Synths in those ERT's to be more support focused
primarily which is why we're removing the ability for them to use guns.
This is done by disabling the allow_gun_usage variable. (The whiteout
deathsquad will not be affected by this change.)

The CLF Synth survivor on LV will also be affected by this change. (The
whiteout deathsquad will not be affected by this change.)

I'll make a separate PR to adjust what equipment synthetics get which
are affected by this change to help facilitate the removal of gun usage.

In the future, we will make a component for Synths that allows the CLF
synth/s specifically to randomly have their "laws" scrambled and switch
between different set/s to bring into the hacked aspect further.

Admins can still enable gun usage as before by setting the VV
allow_gun_usage to be True for their events.

# Explain why it's good for the game
It is very un-fun for both Xeno and Marine to fight a combat synth due
to the power level they are at which includes less stun, no bone breaks,
no pain, and a rather quick repair time. In turn, we want them to be
more on the back-line and assist the actual combat troops with
medication, setting up defenses, maybe handling prisoners and so forth
instead of being a OP killing machine.


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl: Nanu
balance: Removed the ability for Combat Synths in the UPP/CLF ERT to use
guns, as well as the CLF Survivor Synth.
add: Did a few minor tweaks to help fill in some of the space given by
the removal of guns/ammo from the Combat Synths loadout.
/:cl: