Reverts Removal of M4 Armor From FTL (#5834)
# About the pull request

Readds M4 armor to the FTL vendor prep.
<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game

The removal of the M4 armor by PR #5698 was out of scope for the changes
it brought. The M4 was intended as a unique item for RTOs/FTLs and
removing it because "This is to avoid people choosing FTL for the extra
free armor instead of picking Rifleman" is a rather weak reasoning,
there is no data to support the argument that players pick the FTL
solely for free extra armor.

On any given round it is already difficult to fill the slots for FTL.
Removing features that make the role unique is not the way to go about
increasing quality of the players who choose it, if anything it would
have the opposite effect.

# Testing Photographs and Procedure

<details>
<summary>Screenshots & Videos</summary>

</details>


# Changelog

:cl:
balance: Readds M4 armor to FTL vendor as a default item
/:cl: