Fix for crates and traps. (#5835)
# About the pull request

This is essentially a revert of #3075.

# Explain why it's good for the game

Firstly, I can challenge the basic idea of that PR. In my humble
personal experience, when I pry apart crates (read: actually barrels), I
am interested in their metal far more than in anything actually
contained within. On the relatively rarer occasions involving actually
prying crates of supplies for their contents, I am going to move the
leftover planks of wood off the tile anyway - it does not make much
difference whether they are on top or on bottom of the pile of supplies.

But more importantly, I have to question the execution. Messing with
layers of items on the fly is not a good idea. It is even worse idea to
extend it to all the movable atoms instead of simply items. I assume
that the intent is to affect structures, but side effects of that allow
an outright exploit. If you know, you know.

Some of the machinery crates can use a slight adjustment to make them
look neater. I will probably get to that in a next PR once this one
passes.

# Changelog

:cl:
fix: Removed an exploit involving crates.
/:cl: