Dynamic price for omnisentryguns round 3  (#5702)
# About the pull request

just copy of https://github.com/cmss13-devs/cmss13/pull/5226 but gona
place it all here too

Adds dinamic price to sentrguns, starts at 300 and goes up 100 points
after each purchase. Aims to keep the anti spam meshures in place (makes
them stricter in fact), while allowing players to purchase few of them
for reasonable price.



# Explain why it's good for the game

Please before you put angry comments read this.

Omnisentryguns are not used at all, and that is a shame, due to the
changes that were were put in place to prevent sentryspam(they also
weakened the sentrgun a lot along the way, it was not just price change)
, but made them too expensive to buy just one.

This PR changes it by setting the initial price to 300 and it goes up
100 points after each purchase.
This makes it so that no pilot has reason to not buy at least one.
It is balanced around 3 sentrguns for 400 points each, 5 sentrguns 500
points each ( old price) and 11 sentryguns 800 points each(that means
for 8800 points, same as now), and if you try to buy more then current
roundstart maximum of 11, they will cost more then they do now on
avarage.

This should enshure that FTLs and engineers can count with 1- 3
sentryguns each round and plan acordingly, decide if they want some for
FOB, comms and hopfuly have one left to set up frontal outpost.

In terms of effect on xeno gameplay, the sentrygun is quite weak after
the nerfs, alone it can be taken down by basicly any two xenos with just
a bit of efford and fold when t3 or defender attacks it. That means it
serves the intended purpose, covering medics and engineers from skirmish
castes, while slightly helping agains bigger threats while it is part of
defended outpost.

Please mind that the sentryguns can not be moved, so when marines
overextend past it or are forced to retreat from it the sentry is good
as dead.

In terms of pilot gameplay it kind of nerfs experienced pilots who get
loads of calls as they will be expected to spend some points on
sentryguns and have less ammo as resoult and less expirienced pilots who
would not be able to make use of the ammo anyway are given another tool
that is easier to use and makes them contribute to the round.


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
balance: omnisentrygun price down to 300, but it goes up 100 points
after each purchase
/:cl:

---------

Co-authored-by: Zonespace <41448081+Zonespace27@users.noreply.github.com>
Co-authored-by: vincibrv <vojtechvincibr@seznam.cz>