adds stun to exiting maintenance stations (#5822)
# About the pull request

stun when exiting maintenance stations like in
/obj/structure/closet/proc/dump_contents()
<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game

makes it function similarly to other places mobs can hide in
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl: Lok1
add: maintenance stations stun when exiting like closets
/:cl:

---------

Co-authored-by: SabreML <57483089+SabreML@users.noreply.github.com>