Removes an unused variable from code (#5844)
# About the pull request
Removes "ignore_slots" from hive status, as far as I can tell it doesn't
do anything? nowhere else is it mentioned in the code, and the gitlab
archive is the same.


<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
useless vars... bad?
# Testing Photographs and Procedure
<details>
<summary>Testing </summary>

it compiled.

</details>


# Changelog
no player facing changes