Adds Professor DUMMY to the CMO office, accessible by the CMO and the SEA (#5874)
# About the pull request

Adds a new wall cabinet to the CMO's office called the "professor dummy
cabinet". It contains Professor DUMMY and its tablet: it is a tool for
new nurses(!!) and doctors to learn medical treatment, including
surgery.

The locker can only be opened by the CMO or the SEA (not even the CO can
open it). It cannot be flashbanged, melted, or slashed.

There is a hefty warning for everyone when the dummy falls out, in case
someone destroys the closet in a creative way:


![image](https://github.com/cmss13-devs/cmss13/assets/49321394/632c6bff-02dc-428b-9d21-4f5acedd9db8)

The dummy and its tablet dust upon hijack impact, to avoid either side
having a godmoded mob.

# Explain why it's good for the game

It is a tool for new medical personnel to practice on. It can also ease
boredom if the round is going too smoothly for marines.

# Testing Photographs and Procedure

Setting up the tablet and hijacking. (The dummy no longer spawns in the
Morgue but in the CMO office).

<details>
<summary>Screenshots & Videos</summary>


https://github.com/cmss13-devs/cmss13/assets/49321394/563db21d-ebc2-41e0-a8e8-764f9b8aa61b

</details>


# Changelog

:cl:
add: Added a professor dummy cabinet to the CMO office, containing
Professor DUMMY and its control tablet. It is a tool to teach new
medical personnel with. The cabinet is accessible by the CMO and the SEA
only.
/:cl:

---------

Co-authored-by: Drathek <76988376+Drulikar@users.noreply.github.com>