Powernet Fixes (#5891)
# About the pull request

This PR should address some issues with SMES not connecting to the
powernet and reactors that get changed via nightmare inserts throwing
runtimes. It seems that qdeleted reactors were trying to connect to the
powernet in nullspace.

# Explain why it's good for the game

Less runtimes; and hopefully better powernet utilization.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl: Drathek
fix: Fixes some issues with SMES and reactors connecting to the powernet
/:cl: