Fiorina Nightmare Insert (#5857)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

# About the pull request

The Fiorina Prison has seen more and more unrest the last few years,
quelling it down with the dwindling manpower. One such instance, the CMB
picked up a beacon that was sent from the Prison. Thinking it was
another riot, they asked for some Riot Control officers and sent them
with a CMB Deputy to oversee the pacification of the station.

With no greetings at the LZ, they headed to the Armory to speak to
someone in command. What they did not account for was that.. this was no
riot, not anymore at least. The Security Hub is littered with both
prisoner and Security forces corpses. Some sort of green blood and heavy
fortifications. A team member snatched in the darkness and his screams
silencing after a couple of seconds prove that... there's nobody left.

You must survive until the CMB or some other force finds you.. and saves
you before it's too late.

# Explain why it's good for the game

Woohooo nightmare insert baby yeayh


# Testing Photographs and Procedure
<!-- Include any screenshots/videos/debugging steps of the modified code
functioning successfully, ideally including edge cases. -->

<!-- !! If you are modifying sprites, you **must** include one or more
in-game screenshots or videos of the new sprites. !! -->

<details>
<summary>Screenshots & Videos</summary>


![image](https://github.com/cmss13-devs/cmss13/assets/79063506/b3d5cad0-a136-4c5a-bf3e-12a315c5eebd)

</details>


# Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
label your changes in the changelog. Please note that maintainers freely
reserve the right to remove and add tags should they deem it
appropriate. You can attempt to finagle the system all you want, but
it's best to shoot for clear communication right off the bat. -->
<!-- If you add a name after the ':cl', that name will be used in the
changelog. You must add your CKEY after the CL if your GitHub name
doesn't match. Maintainers freely reserve the right to remove and add
tags should they deem it appropriate. -->

:cl:
mapadd: added a new Fiorina nightmare insert with CMB and UA Officers
responding to the distress signal.
/:cl:
<!-- Both :cl:'s are required for the changelog to work! -->

---------

Co-authored-by: Jeff Watchson <johncenatoutourourou@gmail.com>