Fix LateInitialize not occurring for poddoors for relativewall_neighbours (#5892)
# About the pull request

This PR is a followup to #5821 where I forgot to check if LateInitialize
was occurring since it requires a hint to do so. Previously they would
call the proc after a timer, but it made more sense to do so in
LateInitialize. However, I am not able to find anywhere this even
visually has any effect so I'm not sure its necessary for the doors to
be calling `relativewall_neighbours`.

# Explain why it's good for the game

Fixes dead code.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl: Drathek
fix: Fix pod doors no longer updating adjacent tiles
/:cl: