buffs ravager empower (#5890)
# About the pull request
empower range increased to 4

# Explain why it's good for the game
#3729 nerfed base rav too much, rav does not have great armor (only 25)
which means it has to rely alot on shields gained for sustain. it also
increased the max targets from 5 to 6 but didn't give any empower range,
its very rare to actually be able to get the full 6 targets.

# Testing Photographs and Procedure
its a numbers change

# Changelog
:cl:
balance: ravager empower range is now 4 tiles
/:cl: