Fix for empty lists of frequencies. (#5898)
# About the pull request

Followup to #5833. First I thought that I had simply fumbled back then
by fixing things enough that breaking them actually works now and fixing
it is simply a matter of making the list empty itself instead of
replacing itself with a null so that trying to `|=` to it no longer
would runtime.
And then I discovered that somebody once had decided that the syntax to
declare comms machines as universal shall be by giving them empty
frequency lists. Please never do that. It is a pain to track which cases
are intended to be empty because they are meant to be universal and
which are intended to be empty because they will be filled via some
procedure. I _think_ that I sorted them correctly, but if something will
break again, more fixing will need doing.

# Explain why it's good for the game

Is fix.

# Changelog

:cl:
fix: Wiping frequencies off radiotowers no longer breaks them forever.
/:cl: