Dropship doors can now always be broken by queen (#5917)
# About the pull request

This PR is a follow up to
https://github.com/cmss13-devs/cmss13/pull/5808 removing a locked check
to break the doors.

# Explain why it's good for the game

Now regardless of currently locked or open, the queen can break the
doors to prevent further door spamming.

<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl: ihatethisengine2
fix: Dropship door prying can now always be performed by queen even if
open or not locked
/:cl:

---------

Co-authored-by: Drathek <76988376+Drulikar@users.noreply.github.com>