Adds an option to mute announcements while being an observer (#5895)
# About the pull request

Adds a verb called `Hear/Silence Ghost Announcements`, accessible via
the Preferences tab and in Character Setup. It toggles the announcement
audio cue (queen screech, CIC beep, etc.) while the player is being an
**observer**. It does not affect in-character audio.

~~Slight issue with the PR right now is that it makes every current
player's announcements silent by default, it only turns on when creating
new prefs (if I did that correctly). I assume I would need @harryob for
updating our current players' prefs for this?~~

Edit: Updated the savefile with harry's guidance so it is turned on by
default for existing players as well.

# Explain why it's good for the game

Sometimes you tab out of the game for a few minutes, but god forbid you
are trying to watch YouTube or just chilling, because CIC and the Queen
are both spamming their ear-destroying announcements.

I assume this would be nice for staff too who are only online to listen
to ahelps.

# Testing Photographs and Procedure

1. Toggle the preference via the Preferences tab
2. Use the Report: verbs in the Admin tab

# Changelog

:cl:
add: Added an option to hear or silence announcement audio cues (queen
screech, CIC beep, etc.) while being an observer.
/:cl:

---------

Co-authored-by: Drathek <76988376+Drulikar@users.noreply.github.com>
Co-authored-by: SabreML <57483089+SabreML@users.noreply.github.com>