Balance : adding M44 Marksman Speed Loader and SU-6 Smartpistol Magazine to marine squads vendors. (#5908)
# About the pull request

<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
Mainly for consistency and because those ammo are available for free in
any prep.
allowing  marine to buy some if vendor is empty seem fair to me.
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
balance: adding M44 Marksman Speed Loader and SU-6 Smartpistol Magazine
in vendor of (medic,comtech, SL) at a cost of 6 each.
balance: adding M44 Marksman Speed Loader and SU-6 Smartpistol Magazine
in vendor of (rifleman, TL, IO) at a cost of 10 each.
balance: adding VP78 case and SU-6 case in vendor of IO at a cost of 15
each.
/:cl:

---------

Co-authored-by: Julien <jverger.ingx@gmail.com>