Changes to SG vendor add the ability to buy more sidearm ammo type and a smartgun battery. (#5913)
# About the pull request
<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
balance: adding M44 Marksman Speed Loader and SU-6 Smartpistol Magazine
in SG vendor for 10 point each.
balance: adding the ability for to buy smartgun DV9 battery for 15
points in SG vendor.
/:cl:

---------

Co-authored-by: Julien <jverger.ingx@gmail.com>
Co-authored-by: Drathek <76988376+Drulikar@users.noreply.github.com>