Makes 'Join as Xeno' disable admin larva protection (#5934)
# About the pull request

Makes the 'Join as Xeno' button/verb automatically disable admin larva
protection when clicked.

# Explain why it's good for the game

Mostly just a QOL thing. Currently you have to go to the admin tab and
click the verb every time you die in order to actually join the xeno
queue, whereas with this it's tied directly to the usual button.
There shouldn't be any issues with it being clicked accidentally either,
since larva protection can always be re-enabled using the verb as usual.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
admin: Made the 'Join as Xeno' button disable the user's larva
protection when clicked by a moderator.
/:cl: