refactor :  poddoor file and shutters file. (#5886)
# About the pull request
1-added a folder for poddoor and shutters.
2-move those files inside their folders.
3-extracting almayer and two tile children of poddoor and give them
their own files.
<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
refactor: refactored poddoor file and shutters file.
/:cl:

---------

Co-authored-by: Julien <jverger.ingx@gmail.com>