add some balloons for CPR : one is for when it's a success and the other is when it's a failure. (#5918)
# About the pull request
adding balloons to show when you are successful with your CPR or if your
not.

I am not sure about the text.
<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
Give feedback without needing to look at chat box
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
qol: add some balloons for CPR : one is for when it's a success and the
other is when it's a failure.
/:cl:

---------

Co-authored-by: Julien <jverger.ingx@gmail.com>
Co-authored-by: SabreML <57483089+SabreML@users.noreply.github.com>