Tacmap makes no update noise for observers when hear/silence ghost announcement is toggled on (#5990)
# About the pull request

This is a follow-up to #5895 that mutes announcement sounds with a new
preference option.

Tacmaps use a different proc that I missed the first time, this PR mutes
them too.

# Explain why it's good for the game

Same as in #5895, if someone just wants to be online but not get
jumpscared by random noises (e.g. staff or people waiting for
larva/ERT), this helps them.

# Testing Photographs and Procedure

1. Spawn in as CO
2. Spawn in cotablet
3. Mute the CO game instance
4. Join the server with another ckey
5. Observe
6. Mute their ghost announcements
7. Update tacmap as CO
8. Unmute the observer's ghost announcements
9. Update tacmap as CO


# Changelog

:cl:
add: Tacmap announcements are now also muted if you have hear/silence
ghost announcements toggled on.
/:cl:

---------

Co-authored-by: Segrain <Segrain@users.noreply.github.com>