Fortune Changes (#5994)
# Fortune Changes

The purpose of this PR is to add and fix a grammar error I noticed with
a fortune. No mapping changes or other changes have been made.

# Explain why it's good for the game
This update is good for the game because it adds to a world that feels
lived in, adding fortunes that increase the immersion of the game for
the players.

# Testing Photographs and Procedure
<details>
N/A, no screenshots or videos.
</details>


# Changelog
:cl:
add: Added a couple of fortunes in strings/fortunes.txt
spellcheck: fixed a grammar error I found in strings/fortunes.txt
/:cl:

Co-authored-by: DN9123 <dn9123@icloud.com>