Fixes zombies missing claws on transformation (#5998)
# About the pull request
Fixes #5485, added a simple conditional to check if a mob is buckled and
then unbuckle them before transformation. I didn't look into the root of
the problem but this bug seems niche enough that it shouldn't really
effect anything else. If I were to guess, I'm assuming it has to do with
the fact that you can't hold items in your hands while buckled to a bed,
but you can while you're in a chair. Hence why everything works fine
while buckled to a chair.

# Changelog
:cl:
fix: Fixes zombies missing claws on transformation
/:cl: