Allows wardens to heal eachother again at half the health and gives a bigger message (#5982)
# About the pull request
Allows wardens to heal eachother once again, but for half the health
this time.

# Explain why it's good for the game
This was removed because people were dual wardening and keeping
eachother alive very well
this should tackle that and hopefully only allow it to be used on crit /
near crit situations so.
Also gives reju/heal a bigger message so its easier to see if it went
through.

I hate not being able to save another warden because hes of the same
strain
# Testing Photographs and Procedure
i made it so its good
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
balance: Wardens can heal eachother once again, but only half the health
qol: Gives SPAN_XENOHIGHDANGER to heal/reju on positives, since its very
easy to miss if you actually healed soembody.
/:cl:

---------

Co-authored-by: InsaneRed <prodexter31@outlook.comm>
Co-authored-by: Birdtalon <birdtalon@gmail.com>